- Fixed some of my skill issues - Added action buttons - Added placeholders to bossbar progress - Fixed some issues with PTSL & changed some syntax's - Fixed particle data not saving - Unrolledback my changes - Fixed actions with locations - Added Bucket Entity Event & Permission - Fixed Offline Build - Added undo delete action/condition button to actions menu - Made houses not unload if you died in them

+++ paper/src/main/java/com/al3x/housing2/Menus/EditHouseMenu.java
@@ -63,7 +63,9 @@ public class EditHouseMenu extends Menu{
                 .description("Delete this house")
                 .build(), () -> {
                     player.closeInventory();
-                    main.getHousesManager().deleteHouse(house.getHouseUUID());
+                    new ConfirmMenu(player, this, "&cYou want to delete your house?", () -> {
+                        main.getHousesManager().deleteHouse(house.getHouseUUID());
+                    }).open();
                 }
         );
